Added test + logic for user, game both repository and service

+++ FinalAssignmentBE/Models/GameRule.cs
@@ -5,9 +5,9 @@ namespace FinalAssignmentBE.Models;
 public class GameRule
 {
     [Key]
-    public int RuleId { get; set; }
+    public long RuleId { get; set; }
     public int DivisibleNumber { get; set; }
     public string ReplacedWord { get; set; }
-    public int GameId { get; set; }
+    public long GameId { get; set; }
     public Game Game { get; set; }
 }
